refactor(frontend)[OK-325] Comments taken into account
