+++ client/src/config/brushConfig.ts
@@ -1,0 +1,3 @@
+export enum BrushConfig {
+  DECIMATE = 4,
+}
