+++ client/src/config/zoomConfig.ts
@@ -1,4 +1,4 @@
-export enum ZoomOptions {
+export enum ZoomConfig {
   MAX_ZOOM = 8,
   ZOOM_SMOOTHICITY = 0.995, // it should be less than 1 and greater than 0 , 0.999 - the most smooth zoom
 }
