+++ client/src/data/start-page/get-started-data.tsx
