+++ client/src/data/start-page/hexagons-data.ts
