+++ client/src/data/start-page/join-community-data.tsx
