+++ client/src/data/start-page/useful-capabilities-section.ts
