--- client/src/enums/BrushOptions.ts
@@ -1,3 +1,0 @@
-export enum BrushOptions {
-  DECIMATE = 4,
-}
